Deleting hr_payslip_leave
