+++ Py_todo.txt
@@ -1,0 +1,4 @@
+My to do list
+
+1. Restrcit multiple payslip per month for one salary
+2. Adding wage
