+++ hta_custom_hr/models/__init__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
 from . import models
-from . import hr_payslip_leave
+from . import hr_payslip
 from . import hr_payslip_model
 from . import hr_contract
 from . import hr_employee
