+++ hta_custom_hr/models/hr_payslip.py
@@ -1,0 +1,34 @@
+# -*- coding: utf-8 -*-
+
+from odoo import models, fields, api, _
+from odoo.exceptions import UserError
+import datetime
+
+class HrPayslip(models.Model):
+    _inherit = "hr.payslip"
+    
+    def _get_contract_wage(self):
+        self.ensure_one()
+        return self.contract_id.salaire_base
+    
+    slip_month = fields.Char(string='Slip Month', store=True, compute='compute_slip_month')
+    
+    #check unique payslip for employee for a given date
+    """@api.constrains('slip_month', 'employee_id')
+    def _unique_slip_id_per_date(self):
+        for record in self:
+            res = self.env['hr.payslip'].search([('slip_month', '=', record.slip_month), ('employee_id', '=', record.employee_id.id), ])
+            if len(res) > 1:
+                raise ValidationError("You cannot create duplicate payslip for employee {0}. Payslip already exist.".format(record.employee_id.name))
+    """
+    
+    @api.depends('date_from')
+    def compute_slip_month(self):
+        for slip in self:
+            slip.slip_month = slip.date_from.month
+
+class HrPayslipLine(models.Model):
+    _inherit = 'hr.payslip.line'
+    
+    appears_on_paybook = fields.Boolean(related='salary_rule_id.appears_on_paybook', readonly=False)
+    slip_month = fields.Char(related="slip_id.slip_month", string='Slip Month')
