--- hta_custom_hr/models/hr_payslip_leave.py
@@ -1,14 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from odoo import models, fields, api, _
-from odoo.exceptions import UserError
-import datetime
-
-class HrPayslip(models.Model):
-    _inherit = "hr.payslip"
-    
-    def _get_contract_wage(self):
-        self.ensure_one()
-        return self.contract_id.salaire_base
-    
-
